+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+* Create "setViewport" method for browser tabs.
 * Plugins installations are now based on ids.
 * Remove right click method from browser tabs and create options object in the click method.
 * Return string symbols instead of objects when requesting the account symbols.